+++ src/main/java/com/hivemq/bootstrap/netty/ChannelDependencies.java
@@ -162,10 +162,7 @@ public class ChannelDependencies {
     private final ConnackOutboundInterceptorHandler connackOutboundInterceptorHandler;
 
     @NotNull
-    private final PubackInboundInterceptorHandler pubackInboundInterceptorHandler;
-
-    @NotNull
-    private final PubackOutboundInterceptorHandler pubackOutboundInterceptorHandler;
+    private final PubackInterceptorHandler pubackInterceptorHandler;
 
     @Inject
     public ChannelDependencies(
@@ -204,8 +201,7 @@ public class ChannelDependencies {
             @NotNull final PublishOutboundInterceptorHandler publishOutboundInterceptorHandler,
             @NotNull final ConnectInboundInterceptorHandler connectInboundInterceptorHandler,
             @NotNull final ConnackOutboundInterceptorHandler connackOutboundInterceptorHandler,
-            @NotNull final PubackInboundInterceptorHandler pubackInboundInterceptorHandler,
-            @NotNull final PubackOutboundInterceptorHandler pubackOutboundInterceptorHandler) {
+            @NotNull final PubackInterceptorHandler pubackInterceptorHandler) {
 
 
         this.statisticsInitializer = statisticsInitializer;
@@ -243,8 +239,7 @@ public class ChannelDependencies {
         this.publishOutboundInterceptorHandler = publishOutboundInterceptorHandler;
         this.connectInboundInterceptorHandler = connectInboundInterceptorHandler;
         this.connackOutboundInterceptorHandler = connackOutboundInterceptorHandler;
-        this.pubackInboundInterceptorHandler = pubackInboundInterceptorHandler;
-        this.pubackOutboundInterceptorHandler = pubackOutboundInterceptorHandler;
+        this.pubackInterceptorHandler = pubackInterceptorHandler;
     }
 
     @NotNull
@@ -423,12 +418,7 @@ public class ChannelDependencies {
     }
 
     @NotNull
-    public PubackInboundInterceptorHandler getPubackInboundInterceptorHandler() {
-        return pubackInboundInterceptorHandler;
-    }
-
-    @NotNull
-    public PubackOutboundInterceptorHandler getPubackOutboundInterceptorHandler() {
-        return pubackOutboundInterceptorHandler;
+    public PubackInterceptorHandler getPubackInterceptorHandler() {
+        return pubackInterceptorHandler;
     }
 }
