+++ src/main/java/com/hivemq/bootstrap/netty/initializer/AbstractChannelInitializer.java
@@ -91,9 +91,7 @@ public abstract class AbstractChannelInitializer extends ChannelInitializer<Chan
         ch.pipeline().addLast(CONNECT_INBOUND_INTERCEPTOR_HANDLER, channelDependencies.getConnectInboundInterceptorHandler());
         ch.pipeline().addLast(CLIENT_LIFECYCLE_EVENT_HANDLER, channelDependencies.getClientLifecycleEventHandler());
 
-        ch.pipeline()
-                .addLast(PUBACK_INBOUND_INTERCEPTOR_HANDLER, channelDependencies.getPubackInboundInterceptorHandler());
-        ch.pipeline().addLast(PUBACK_OUTBOUND_INTERCEPTOR_HANDLER, channelDependencies.getPubackOutboundInterceptorHandler());
+        ch.pipeline().addLast(PUBACK_INTERCEPTOR_HANDLER, channelDependencies.getPubackInterceptorHandler());
         ch.pipeline().addLast(PUBLISH_OUTBOUND_INTERCEPTOR_HANDLER, channelDependencies.getPublishOutboundInterceptorHandler());
 
         ch.pipeline().addLast(LISTENER_ATTRIBUTE_ADDER, channelDependencies.getListenerAttributeAdderFactory().get(listener));
